Refactored init function and location methods
