+++ src/utils/index.ts
@@ -1,7 +1,6 @@
 export * from './frequency';
-export { default as hasStartedLocationUpdates } from './hasStartedLocationUpdates';
 export * from './measureMeters';
 export * from './notifications';
 export * from './permissions';
-export { default as startLocationUpdates } from './startLocationUpdates';
+export * from './location';
 export * from './task';
